Switch to new source for zinit
The old repo for zinit was deleted, as its author started a new
project instead, and as a result I couldn't update it. I'm thus switching
to the continuation in the zdharma-continuum repo, as it's the easiest
to switch to, and doesn't require me to change anything from the rest
of my zshrc.